Refactor into modular package
