+++ app.py
@@ -1,929 +1,8 @@
-from flask import (
-    Flask,
-    render_template,
-    request,
-    make_response,
-    session,
-    redirect,
-    url_for,
-    has_request_context,
-    send_from_directory,
-)
+from stockapp import create_app
 import os
-import requests
-from babel import Locale
-from babel.numbers import format_currency, format_decimal
-from babel.dates import format_datetime
-import csv
-import io
-from fpdf import FPDF
-from flask_wtf import CSRFProtect
-from flask_sqlalchemy import SQLAlchemy
-from flask_login import (
-    LoginManager,
-    login_user,
-    logout_user,
-    login_required,
-    current_user,
-    UserMixin,
-)
-from werkzeug.security import generate_password_hash, check_password_hash
-from datetime import datetime, timedelta
-import smtplib
-from email.mime.text import MIMEText
-from apscheduler.schedulers.background import BackgroundScheduler
-import atexit
-import secrets
 
-app = Flask(__name__)
-# Load sensitive configuration from environment variables
-app.config["SECRET_KEY"] = os.environ.get("SECRET_KEY", "change_this_secret")
+app = create_app()
 
-# Use DATABASE_URL if provided (e.g. when deployed on Render), otherwise
-# default to a local SQLite database.
-db_url = os.environ.get("DATABASE_URL", "sqlite:///app.db")
-if db_url.startswith("postgres://"):
-    db_url = db_url.replace("postgres://", "postgresql://", 1)
-
-app.config["SQLALCHEMY_DATABASE_URI"] = db_url
-app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-app.config["SMTP_SERVER"] = os.environ.get("SMTP_SERVER", "smtp.example.com")
-app.config["SMTP_PORT"] = int(os.environ.get("SMTP_PORT", 587))
-app.config["SMTP_USERNAME"] = os.environ.get("SMTP_USERNAME", "user@example.com")
-app.config["SMTP_PASSWORD"] = os.environ.get("SMTP_PASSWORD", "password")
-
-db = SQLAlchemy(app)
-csrf = CSRFProtect(app)
-
-
-@app.route("/service-worker.js")
-def service_worker():
-    """Serve the service worker file with correct scope."""
-    return app.send_static_file("service-worker.js")
-
-login_manager = LoginManager(app)
-login_manager.login_view = "login"
-
-@app.context_processor
-def inject_csrf_token():
-    from flask_wtf.csrf import generate_csrf
-    return dict(csrf_token=generate_csrf)
-
-API_KEY = os.environ.get("API_KEY")
-if not API_KEY:
-    raise RuntimeError("API_KEY environment variable not set")
-
-# Threshold for triggering a P/E ratio alert
-ALERT_PE_THRESHOLD = 30
-
-
-def get_locale():
-    """Determine best match locale from the request."""
-    if has_request_context():
-        loc = request.accept_languages.best or "en_US"
-        try:
-            return str(Locale.parse(loc))
-        except Exception:
-            return "en_US"
-    return "en_US"
-
-
-class User(db.Model, UserMixin):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(150), unique=True, nullable=False)
-    password_hash = db.Column(db.String(150), nullable=False)
-    email = db.Column(db.String(150), unique=True, nullable=True)
-    is_verified = db.Column(db.Boolean, default=False)
-    verification_token = db.Column(db.String(100), unique=True)
-    reset_token = db.Column(db.String(100), unique=True)
-    alert_frequency = db.Column(db.Integer, default=24)
-    last_alert_time = db.Column(db.DateTime, default=datetime.utcnow)
-
-
-class WatchlistItem(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    symbol = db.Column(db.String(10), nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-    pe_threshold = db.Column(db.Float, default=ALERT_PE_THRESHOLD)
-
-
-class FavoriteTicker(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    symbol = db.Column(db.String(10), nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-
-
-class Alert(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    symbol = db.Column(db.String(10))
-    message = db.Column(db.String(200))
-    timestamp = db.Column(db.DateTime, default=datetime.utcnow)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-
-
-class History(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    symbol = db.Column(db.String(10))
-    timestamp = db.Column(db.DateTime, default=datetime.utcnow)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-
-
-@login_manager.user_loader
-def load_user(user_id):
-    return User.query.get(int(user_id))
-
-
-def send_email(to, subject, body):
-    smtp_server = app.config.get("SMTP_SERVER")
-    smtp_port = app.config.get("SMTP_PORT", 587)
-    smtp_user = app.config.get("SMTP_USERNAME")
-    smtp_pass = app.config.get("SMTP_PASSWORD")
-    if not all([smtp_server, smtp_user, smtp_pass]):
-        print("Email configuration incomplete")
-        return
-
-    msg = MIMEText(body)
-    msg["Subject"] = subject
-    msg["From"] = smtp_user
-    msg["To"] = to
-    try:
-        with smtplib.SMTP(smtp_server, smtp_port) as server:
-            server.starttls()
-            server.login(smtp_user, smtp_pass)
-            server.send_message(msg)
-    except Exception as e:
-        print(f"Email error: {e}")
-
-
-def get_historical_prices(symbol, days=30):
-    """Fetch historical closing prices for the given symbol."""
-    url = (
-        f"https://financialmodelingprep.com/api/v3/historical-price-full/{symbol}?"
-        f"serietype=line&timeseries={days}&apikey={API_KEY}"
-    )
-    try:
-        response = requests.get(url, timeout=10)
-        data = response.json()
-        historical = data.get("historical", [])
-        dates = [item.get("date") for item in historical][::-1]
-        prices = [item.get("close") for item in historical][::-1]
-        return dates, prices
-    except Exception as e:
-        print(f"Historical API error: {e}")
-        return [], []
-
-def format_market_cap(value, currency):
-    if value is None:
-        return "N/A"
-    suffix = ""
-    if value >= 1_000_000_000_000:
-        value /= 1_000_000_000_000
-        suffix = "T"
-    elif value >= 1_000_000_000:
-        value /= 1_000_000_000
-        suffix = "B"
-    elif value >= 1_000_000:
-        value /= 1_000_000
-        suffix = "M"
-    formatted = format_currency(value, currency, locale=get_locale())
-    return f"{formatted}{suffix}"
-
-def get_stock_data(symbol):
-    quote_url = f"https://financialmodelingprep.com/api/v3/quote/{symbol}?apikey={API_KEY}"
-    profile_url = f"https://financialmodelingprep.com/api/v3/profile/{symbol}?apikey={API_KEY}"
-    ratios_url = f"https://financialmodelingprep.com/api/v3/ratios-ttm/{symbol}?apikey={API_KEY}"
-    rating_url = f"https://financialmodelingprep.com/api/v3/rating/{symbol}?apikey={API_KEY}"
-    growth_url = f"https://financialmodelingprep.com/api/v3/financial-growth/{symbol}?limit=1&apikey={API_KEY}"
-
-    try:
-        # Get quote data
-        quote_response = requests.get(quote_url, timeout=10)
-        quote_data = quote_response.json()
-        if not isinstance(quote_data, list) or len(quote_data) == 0:
-            return None, None, None, None, None, None, None, None, None, None
-        quote = quote_data[0]
-
-        # Get profile data
-        profile_response = requests.get(profile_url, timeout=10)
-        profile_data = profile_response.json()
-        profile = profile_data[0] if isinstance(profile_data, list) and len(profile_data) > 0 else {}
-
-        ratio_response = requests.get(ratios_url, timeout=10)
-        ratio_data = ratio_response.json()
-        debt_to_equity = pb_ratio = roe = roa = profit_margin = dividend_yield = None
-        price_to_sales = None
-        if isinstance(ratio_data, list) and len(ratio_data) > 0:
-            r = ratio_data[0]
-            debt_to_equity = r.get("debtEquityRatioTTM")
-            pb_ratio = r.get("priceToBookRatioTTM")
-            roe = r.get("returnOnEquityTTM")
-            roa = r.get("returnOnAssetsTTM")
-            profit_margin = r.get("netProfitMarginTTM")
-            dividend_yield = r.get("dividendYielTTM") or r.get("dividendYieldTTM")
-            price_to_sales = r.get("priceToSalesRatioTTM")
-
-        rating_response = requests.get(rating_url, timeout=10)
-        rating_data = rating_response.json()
-        analyst_rating = None
-        if isinstance(rating_data, list) and len(rating_data) > 0:
-            analyst_rating = rating_data[0].get("ratingRecommendation") or rating_data[0].get("rating")
-
-        growth_response = requests.get(growth_url, timeout=10)
-        growth_data = growth_response.json()
-        earnings_growth = None
-        if isinstance(growth_data, list) and len(growth_data) > 0:
-            earnings_growth = growth_data[0].get("growthEPS") or growth_data[0].get("epsgrowth")
-
-        name = profile.get("companyName", "")
-        logo_url = profile.get("image", "")
-        sector = profile.get("sector", "")
-        industry = profile.get("industry", "")
-        exchange = profile.get("exchangeShortName", "")
-        currency = profile.get("currency", "USD")
-
-        price = quote.get("price")
-        eps = quote.get("eps")
-        market_cap = format_market_cap(quote.get("marketCap"), currency)
-
-        forward_pe = None
-        if price is not None and eps is not None:
-            growth = None
-            try:
-                growth = float(earnings_growth) if earnings_growth is not None else None
-            except (TypeError, ValueError):
-                growth = None
-            if growth is not None:
-                try:
-                    forward_eps = eps * (1 + growth)
-                    forward_pe = round(price / forward_eps, 2) if forward_eps else None
-                except Exception:
-                    forward_pe = None
-
-
-        return (
-            name,
-            logo_url,
-            sector,
-            industry,
-            exchange,
-            currency,
-            price,
-            eps,
-            market_cap,
-            debt_to_equity,
-            pb_ratio,
-            roe,
-            roa,
-            profit_margin,
-            analyst_rating,
-            dividend_yield,
-            earnings_growth,
-            forward_pe,
-            price_to_sales,
-        )
-    except Exception as e:
-        print(f"API error: {e}")
-        return (
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-        )
-
-
-def check_watchlists():
-    with app.app_context():
-        now = datetime.utcnow()
-        users = User.query.all()
-        for user in users:
-            if not user.email:
-                continue
-            freq = user.alert_frequency or 24
-            last = user.last_alert_time or datetime.min
-            if now - last < timedelta(hours=freq):
-                continue
-            items = WatchlistItem.query.filter_by(user_id=user.id).all()
-            for item in items:
-                (
-                    _name,
-                    _logo_url,
-                    _sector,
-                    _industry,
-                    _exchange,
-                    currency,
-                    price,
-                    eps,
-                    _market_cap,
-                    _debt_to_equity,
-                    *_rest,
-                ) = get_stock_data(item.symbol)
-                if price is not None and eps:
-                    pe_ratio = round(price / eps, 2)
-                    threshold = item.pe_threshold or ALERT_PE_THRESHOLD
-                    if pe_ratio > threshold:
-                        msg = (
-                            f"{item.symbol} P/E ratio {pe_ratio} exceeds threshold {threshold}"
-                        )
-                        send_email(user.email, "P/E Ratio Alert", msg)
-                        db.session.add(
-                            Alert(symbol=item.symbol, message=msg, user_id=user.id)
-                        )
-            user.last_alert_time = now
-        db.session.commit()
-
-@app.route("/", methods=["GET", "POST"])
-def index():
-    symbol = ""
-    price = eps = pe_ratio = valuation = company_name = logo_url = market_cap = sector = industry = exchange = currency = debt_to_equity = None
-    pb_ratio = roe = roa = profit_margin = analyst_rating = dividend_yield = earnings_growth = None
-    forward_pe = price_to_sales = None
-    error_message = alert_message = None
-    history_dates = history_prices = []
-
-    # Ticker may be provided via query string to allow GET requests without CSRF
-    if request.method == "GET":
-        symbol = request.args.get("ticker", "").upper() or symbol
-
-    if current_user.is_authenticated:
-        history_entries = (
-            History.query.filter_by(user_id=current_user.id)
-            .order_by(History.timestamp.desc())
-            .limit(10)
-            .all()
-        )
-        history = [h.symbol for h in history_entries]
-    else:
-        history = session.get("history", [])
-
-    # Handle form submissions as well as ticker provided via GET
-    if request.method == "POST" or symbol:
-        if request.method == "POST":
-            symbol = request.form["ticker"].upper()
-        try:
-            (
-                company_name,
-                logo_url,
-                sector,
-                industry,
-                exchange,
-                currency,
-                price,
-                eps,
-                market_cap,
-                debt_to_equity,
-                pb_ratio,
-                roe,
-                roa,
-                profit_margin,
-                analyst_rating,
-                dividend_yield,
-                earnings_growth,
-                forward_pe,
-                price_to_sales,
-            ) = get_stock_data(symbol)
-
-            history_dates, history_prices = get_historical_prices(symbol, days=90)
-
-            if price is not None and eps:
-                pe_ratio_val = round(price / eps, 2)
-                pe_ratio = format_decimal(pe_ratio_val, locale=get_locale())
-                if pe_ratio_val < 15:
-                    valuation = "Undervalued?"
-                elif pe_ratio_val > 25:
-                    valuation = "Overvalued?"
-                else:
-                    valuation = "Fairly Valued"
-                threshold = ALERT_PE_THRESHOLD
-                if current_user.is_authenticated:
-                    item = WatchlistItem.query.filter_by(
-                        user_id=current_user.id, symbol=symbol
-                    ).first()
-                    if item and item.pe_threshold is not None:
-                        threshold = item.pe_threshold
-                if pe_ratio_val > threshold:
-                    alert_message = (
-                        f"P/E ratio {pe_ratio_val} exceeds threshold of {threshold}"
-                    )
-                    if current_user.is_authenticated:
-                        db.session.add(
-                            Alert(
-                                symbol=symbol,
-                                message=alert_message,
-                                user_id=current_user.id,
-                            )
-                        )
-                        db.session.commit()
-            elif price is None or eps is None:
-                error_message = "Price or EPS data is missing."
-            if debt_to_equity is not None:
-                debt_to_equity = format_decimal(round(debt_to_equity, 2), locale=get_locale())
-            if pb_ratio is not None:
-                pb_ratio = format_decimal(round(pb_ratio, 2), locale=get_locale())
-            if roe is not None:
-                roe = format_decimal(round(roe * 100, 2), locale=get_locale())
-            if roa is not None:
-                roa = format_decimal(round(roa * 100, 2), locale=get_locale())
-            if profit_margin is not None:
-                profit_margin = format_decimal(round(profit_margin * 100, 2), locale=get_locale())
-            if dividend_yield is not None:
-                dividend_yield = format_decimal(round(dividend_yield * 100, 2), locale=get_locale())
-            if earnings_growth is not None:
-                earnings_growth = format_decimal(round(earnings_growth * 100, 2), locale=get_locale())
-            if forward_pe is not None:
-                forward_pe = format_decimal(round(forward_pe, 2), locale=get_locale())
-            if price_to_sales is not None:
-                price_to_sales = format_decimal(round(price_to_sales, 2), locale=get_locale())
-            if price is not None:
-                price = format_currency(price, currency, locale=get_locale())
-            if eps is not None:
-                eps = format_currency(eps, currency, locale=get_locale())
-        
-            # update history
-            if symbol:
-                if current_user.is_authenticated:
-                    db.session.add(History(symbol=symbol, user_id=current_user.id))
-                    db.session.commit()
-                    history.insert(0, symbol)
-                    history = history[:10]
-                else:
-                    if symbol in history:
-                        history.remove(symbol)
-                    history.insert(0, symbol)
-                    history = history[:10]
-                    session["history"] = history
-
-        except Exception as e:
-            error_message = str(e)
-
-    return render_template(
-        "index.html",
-        symbol=symbol,
-        price=price,
-        eps=eps,
-        pe_ratio=pe_ratio,
-        valuation=valuation,
-        company_name=company_name,
-        logo_url=logo_url,
-        market_cap=market_cap,
-        sector=sector,
-        industry=industry,
-        exchange=exchange,
-        currency=currency,
-        debt_to_equity=debt_to_equity,
-        pb_ratio=pb_ratio,
-        roe=roe,
-        roa=roa,
-        profit_margin=profit_margin,
-        analyst_rating=analyst_rating,
-        dividend_yield=dividend_yield,
-        earnings_growth=earnings_growth,
-        forward_pe=forward_pe,
-        price_to_sales=price_to_sales,
-        error_message=error_message,
-        alert_message=alert_message,
-        history_dates=history_dates,
-        history_prices=history_prices,
-        history=history,
-    )
-
-
-@app.route("/clear_history")
-def clear_history():
-    if current_user.is_authenticated:
-        History.query.filter_by(user_id=current_user.id).delete()
-        db.session.commit()
-    else:
-        session.pop("history", None)
-    return redirect(url_for("index"))
-
-
-@app.route("/download")
-def download():
-    symbol = request.args.get("symbol", "").upper()
-    fmt = request.args.get("format", "csv").lower()
-    if not symbol:
-        return "Symbol missing", 400
-
-    (
-        company_name,
-        _logo_url,
-        sector,
-        industry,
-        exchange,
-        currency,
-        price,
-        eps,
-        market_cap,
-        debt_to_equity,
-        pb_ratio,
-        roe,
-        roa,
-        profit_margin,
-        analyst_rating,
-        dividend_yield,
-        earnings_growth,
-        forward_pe,
-        price_to_sales,
-    ) = get_stock_data(symbol)
-
-    if price is not None and eps:
-        pe_ratio_val = round(price / eps, 2)
-        if pe_ratio_val < 15:
-            valuation = "Undervalued?"
-        elif pe_ratio_val > 25:
-            valuation = "Overvalued?"
-        else:
-            valuation = "Fairly Valued"
-        pe_ratio = format_decimal(pe_ratio_val, locale=get_locale())
-    else:
-        pe_ratio = valuation = "N/A"
-
-    if debt_to_equity is not None:
-        debt_to_equity = format_decimal(round(debt_to_equity, 2), locale=get_locale())
-    if pb_ratio is not None:
-        pb_ratio = format_decimal(round(pb_ratio, 2), locale=get_locale())
-    if roe is not None:
-        roe = format_decimal(round(roe * 100, 2), locale=get_locale())
-    if roa is not None:
-        roa = format_decimal(round(roa * 100, 2), locale=get_locale())
-    if profit_margin is not None:
-        profit_margin = format_decimal(round(profit_margin * 100, 2), locale=get_locale())
-    if dividend_yield is not None:
-        dividend_yield = format_decimal(round(dividend_yield * 100, 2), locale=get_locale())
-    if earnings_growth is not None:
-        earnings_growth = format_decimal(round(earnings_growth * 100, 2), locale=get_locale())
-    if forward_pe is not None:
-        forward_pe = format_decimal(round(forward_pe, 2), locale=get_locale())
-    if price_to_sales is not None:
-        price_to_sales = format_decimal(round(price_to_sales, 2), locale=get_locale())
-
-    if price is not None:
-        price = format_currency(price, currency, locale=get_locale())
-    if eps is not None:
-        eps = format_currency(eps, currency, locale=get_locale())
-
-    if fmt == "csv":
-        output = io.StringIO()
-        writer = csv.writer(output)
-        writer.writerow(
-            [
-                "Company Name",
-                "Symbol",
-                "Price",
-                "EPS",
-                "P/E Ratio",
-                "Valuation",
-                "Market Cap",
-                "Debt/Equity",
-                "P/B",
-                "ROE %",
-                "ROA %",
-                "Profit Margin %",
-                "Analyst Rating",
-                "Dividend Yield %",
-                "Earnings Growth %",
-                "Forward P/E",
-                "P/S Ratio",
-                "Sector",
-                "Industry",
-                "Exchange",
-                "Currency",
-            ]
-        )
-        writer.writerow([
-            company_name,
-            symbol,
-            price,
-            eps,
-            pe_ratio,
-            valuation,
-            market_cap,
-            debt_to_equity,
-            pb_ratio,
-            roe,
-            roa,
-            profit_margin,
-            analyst_rating,
-            dividend_yield,
-            earnings_growth,
-            forward_pe,
-            price_to_sales,
-            sector,
-            industry,
-            exchange,
-            currency,
-        ])
-        response = make_response(output.getvalue())
-        response.headers["Content-Disposition"] = f"attachment; filename={symbol}_data.csv"
-        response.headers["Content-Type"] = "text/csv"
-        return response
-    elif fmt == "pdf":
-        pdf = FPDF()
-        pdf.add_page()
-        pdf.set_font("Arial", size=12)
-        pdf.cell(0, 10, txt=f"Stock Data for {symbol}", ln=1)
-        fields = [
-            ("Company Name", company_name),
-            ("Price", price),
-            ("EPS", eps),
-            ("P/E Ratio", pe_ratio),
-            ("Valuation", valuation),
-            ("Market Cap", market_cap),
-            ("Debt/Equity", debt_to_equity),
-            ("P/B", pb_ratio),
-            ("ROE %", roe),
-            ("ROA %", roa),
-            ("Profit Margin %", profit_margin),
-            ("Analyst Rating", analyst_rating),
-            ("Dividend Yield %", dividend_yield),
-            ("Earnings Growth %", earnings_growth),
-            ("Forward P/E", forward_pe),
-            ("P/S Ratio", price_to_sales),
-            ("Sector", sector),
-            ("Industry", industry),
-            ("Exchange", exchange),
-            ("Currency", currency),
-        ]
-        for label, value in fields:
-            pdf.cell(0, 10, txt=f"{label}: {value}", ln=1)
-        pdf_output = pdf.output(dest="S").encode("latin-1")
-        response = make_response(pdf_output)
-        response.headers["Content-Disposition"] = f"attachment; filename={symbol}_data.pdf"
-        response.headers["Content-Type"] = "application/pdf"
-        return response
-    else:
-        return "Invalid format", 400
-
-
-@app.route("/signup", methods=["GET", "POST"])
-def signup():
-    error = None
-    message = None
-    if request.method == "POST":
-        username = request.form["username"]
-        email = request.form.get("email")
-        password = request.form["password"]
-        if User.query.filter_by(username=username).first():
-            error = "Username already exists"
-        else:
-            token = secrets.token_urlsafe(16)
-            user = User(
-                username=username,
-                password_hash=generate_password_hash(password),
-                email=email,
-                verification_token=token,
-            )
-            db.session.add(user)
-            db.session.commit()
-            verify_link = url_for("verify_email", token=token, _external=True)
-            send_email(
-                email,
-                "Verify your account",
-                f"Please click the link to verify your account: {verify_link}",
-            )
-            message = "Check your email to verify your account."
-    return render_template("signup.html", error=error, message=message)
-
-
-@app.route("/verify/<token>")
-def verify_email(token):
-    user = User.query.filter_by(verification_token=token).first()
-    if user:
-        user.is_verified = True
-        user.verification_token = None
-        db.session.commit()
-        msg = "Email verified. You can now log in."
-    else:
-        msg = "Invalid or expired verification link."
-    return render_template("verify_email.html", message=msg)
-
-
-@app.route("/login", methods=["GET", "POST"])
-def login():
-    error = None
-    message = None
-    if request.method == "POST":
-        username = request.form["username"]
-        password = request.form["password"]
-        user = User.query.filter_by(username=username).first()
-        if user and check_password_hash(user.password_hash, password):
-            if not user.is_verified:
-                message = "Please verify your email before logging in."
-            else:
-                login_user(user)
-                return redirect(url_for("index"))
-        else:
-            error = "Invalid credentials"
-    return render_template("login.html", error=error, message=message)
-
-
-@app.route("/forgot_password", methods=["GET", "POST"])
-def forgot_password():
-    error = None
-    message = None
-    if request.method == "POST":
-        email = request.form.get("email")
-        user = User.query.filter_by(email=email).first()
-        if user:
-            token = secrets.token_urlsafe(16)
-            user.reset_token = token
-            db.session.commit()
-            reset_link = url_for("reset_password", token=token, _external=True)
-            send_email(
-                email,
-                "Password reset",
-                f"Click the link to reset your password: {reset_link}",
-            )
-            message = "Check your email for a password reset link."
-        else:
-            error = "Email not found"
-    return render_template("forgot_password.html", error=error, message=message)
-
-
-@app.route("/reset_password/<token>", methods=["GET", "POST"])
-def reset_password(token):
-    error = None
-    message = None
-    user = User.query.filter_by(reset_token=token).first()
-    if not user:
-        error = "Invalid or expired token"
-        return render_template("reset_password.html", error=error)
-    if request.method == "POST":
-        password = request.form.get("password")
-        if password:
-            user.password_hash = generate_password_hash(password)
-            user.reset_token = None
-            db.session.commit()
-            message = "Password updated. You can now log in."
-        else:
-            error = "Password required"
-    return render_template("reset_password.html", error=error, message=message)
-
-
-@app.route("/logout")
-@login_required
-def logout():
-    logout_user()
-    return redirect(url_for("index"))
-
-
-@app.route("/watchlist", methods=["GET", "POST"])
-@login_required
-def watchlist():
-    if request.method == "POST":
-        if request.form.get("item_id"):
-            item_id = int(request.form["item_id"])
-            threshold = request.form.get("threshold", type=float)
-            item = WatchlistItem.query.get_or_404(item_id)
-            if item.user_id == current_user.id:
-                item.pe_threshold = threshold or ALERT_PE_THRESHOLD
-                db.session.commit()
-        else:
-            symbol = request.form["symbol"].upper()
-            threshold = request.form.get("threshold", type=float) or ALERT_PE_THRESHOLD
-            if not WatchlistItem.query.filter_by(user_id=current_user.id, symbol=symbol).first():
-                db.session.add(
-                    WatchlistItem(
-                        symbol=symbol,
-                        user_id=current_user.id,
-                        pe_threshold=threshold,
-                    )
-                )
-                db.session.commit()
-    items = WatchlistItem.query.filter_by(user_id=current_user.id).all()
-    return render_template(
-        "watchlist.html", items=items, default_threshold=ALERT_PE_THRESHOLD
-    )
-
-
-@app.route("/watchlist/delete/<int:item_id>")
-@login_required
-def delete_watchlist_item(item_id):
-    item = WatchlistItem.query.get_or_404(item_id)
-    if item.user_id == current_user.id:
-        db.session.delete(item)
-        db.session.commit()
-    return redirect(url_for("watchlist"))
-
-
-@app.route("/add_watchlist/<symbol>")
-@login_required
-def add_watchlist(symbol):
-    symbol = symbol.upper()
-    if not WatchlistItem.query.filter_by(user_id=current_user.id, symbol=symbol).first():
-        db.session.add(
-            WatchlistItem(
-                symbol=symbol,
-                user_id=current_user.id,
-                pe_threshold=ALERT_PE_THRESHOLD,
-            )
-        )
-        db.session.commit()
-    return redirect(url_for("index", ticker=symbol))
-
-
-@app.route("/favorites", methods=["GET", "POST"])
-@login_required
-def favorites():
-    if request.method == "POST":
-        symbol = request.form["symbol"].upper()
-        if not FavoriteTicker.query.filter_by(user_id=current_user.id, symbol=symbol).first():
-            db.session.add(FavoriteTicker(symbol=symbol, user_id=current_user.id))
-            db.session.commit()
-    items = FavoriteTicker.query.filter_by(user_id=current_user.id).all()
-    return render_template("favorites.html", items=items)
-
-
-@app.route("/favorites/delete/<int:item_id>")
-@login_required
-def delete_favorite(item_id):
-    item = FavoriteTicker.query.get_or_404(item_id)
-    if item.user_id == current_user.id:
-        db.session.delete(item)
-        db.session.commit()
-    return redirect(url_for("favorites"))
-
-
-@app.route("/add_favorite/<symbol>")
-@login_required
-def add_favorite(symbol):
-    symbol = symbol.upper()
-    if not FavoriteTicker.query.filter_by(user_id=current_user.id, symbol=symbol).first():
-        db.session.add(FavoriteTicker(symbol=symbol, user_id=current_user.id))
-        db.session.commit()
-    return redirect(url_for("index", ticker=symbol))
-
-
-@app.route("/settings", methods=["GET", "POST"])
-@login_required
-def settings():
-    if request.method == "POST":
-        freq = request.form.get("frequency", type=int)
-        if freq and freq > 0:
-            current_user.alert_frequency = freq
-            db.session.commit()
-    return render_template("settings.html", frequency=current_user.alert_frequency)
-
-
-@app.route("/alerts")
-@login_required
-def alerts():
-    entries = (
-        Alert.query.filter_by(user_id=current_user.id)
-        .order_by(Alert.timestamp.desc())
-        .all()
-    )
-    return render_template("alerts.html", alerts=entries)
-
-
-@app.route("/clear_alerts")
-@login_required
-def clear_alerts():
-    Alert.query.filter_by(user_id=current_user.id).delete()
-    db.session.commit()
-    return redirect(url_for("alerts"))
-
-
-@app.route("/export_history")
-@login_required
-def export_history():
-    entries = (
-        History.query.filter_by(user_id=current_user.id)
-        .order_by(History.timestamp.desc())
-        .all()
-    )
-    output = io.StringIO()
-    writer = csv.writer(output)
-    writer.writerow(["Symbol", "Timestamp"])
-    for e in entries:
-        timestamp = format_datetime(e.timestamp, locale=get_locale())
-        writer.writerow([e.symbol, timestamp])
-    response = make_response(output.getvalue())
-    response.headers["Content-Disposition"] = "attachment; filename=history.csv"
-    response.headers["Content-Type"] = "text/csv"
-    return response
-
-if __name__ == "__main__":
-    with app.app_context():
-        db.create_all()
-    scheduler = BackgroundScheduler()
-    scheduler.add_job(check_watchlists, "interval", hours=1)
-    scheduler.start()
-    atexit.register(lambda: scheduler.shutdown())
-    debug_mode = os.environ.get("FLASK_DEBUG", "0").lower() in ["1", "true", "yes"]
-    app.run(host="0.0.0.0", port=5000, debug=debug_mode)
+if __name__ == '__main__':
+    debug_mode = os.environ.get('FLASK_DEBUG', '0').lower() in ['1', 'true', 'yes']
+    app.run(host='0.0.0.0', port=5000, debug=debug_mode)
