+++ stockapp/auth/__init__.py
@@ -1,0 +1,3 @@
+from .routes import auth_bp
+
+__all__ = ['auth_bp']
