+++ stockapp/main/__init__.py
@@ -1,0 +1,3 @@
+from .routes import main_bp
+
+__all__ = ['main_bp']
