+++ stockapp/watchlists/__init__.py
@@ -1,0 +1,3 @@
+from .routes import watch_bp
+
+__all__ = ['watch_bp']
